Quick & Dirty way to bootstrap new account, this needs testing
This is a quick and dirty way to bootstrap a new account, it returns the public key of the project and private key of the project along with the project id. This should give the user everything need to start using the application.

Kinda a chicken and egg thing here but this should be a path forward that could solve that first account creation.